+++ src/dtypes.rs
@@ -1,7 +1,19 @@
-use std::{
-    fmt::{self, Display},
-    ops::{Index, IndexMut},
-};
+//! # Myco Data Types
+//! 
+//! This module contains the core data types used throughout the Myco library.
+//! 
+//! The main types defined here are:
+//! - `Key`: Cryptographic keys used for encryption and PRF operations
+//! - `Path`: Binary paths used in the tree data structure
+//! - `Bucket`: Storage units containing encrypted message blocks
+//! - `Metadata`: Associated metadata for message blocks including paths and timestamps
+//! 
+//! These types form the foundation for Myco's metadata-hiding encrypted messaging system,
+//! enabling secure communication while obscuring patterns of interaction between users.
+//! The types are designed to work together to implement the ORAM-inspired data structure
+//! that provides efficient read/write operations while maintaining strong privacy guarantees.
+
+use std::ops::{Index, IndexMut};
 
 use rand::{seq::SliceRandom, Rng, RngCore, SeedableRng};
 use rand_chacha::ChaCha20Rng;
@@ -12,25 +24,31 @@ use crate::{tree::TreeValue, BLOCK_SIZE, D, LAMBDA, Z};
 pub(crate) type Timestamp = u64;
 
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
+/// Metadata for a message block, including the path, key, and timestamp
 pub struct Metadata(Vec<(Path, Key, Timestamp)>);
 
 impl Metadata {
+    /// Create a new Metadata instance with a single entry
     pub fn new(path: Path, key: Key, timestamp: Timestamp) -> Self {
         Metadata(vec![(path, key, timestamp)])
     }
 
+    /// Get the number of entries in the Metadata
     pub fn len(&self) -> usize {
         self.0.len()
     }
 
+    /// Add a new entry to the Metadata
     pub fn push(&mut self, path: Path, key: Key, timestamp: Timestamp) {
         self.0.push((path, key, timestamp));
     }
 
+    /// Get the entry at a specific index
     pub fn get(&self, index: usize) -> Option<&(Path, Key, Timestamp)> {
         self.0.get(index)
     }
 
+    /// Shuffle the entries in the Metadata using a random number generator
     pub fn shuffle<R: RngCore + Rng>(&mut self, rng: &mut R) {
         self.0.shuffle(rng);
     }
@@ -49,8 +67,11 @@ impl TreeValue for Metadata {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
+/// An enum representing the direction of a path in the tree
 pub enum Direction {
+    /// Left direction
     Left,
+    /// Right direction
     Right,
 }
 
@@ -74,25 +95,31 @@ impl From<u8> for Direction {
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+/// A binary path in the tree, represented as a vector of directions
 pub struct Path(pub Vec<Direction>);
 
 impl Path {
+    /// Create a new Path instance with a given vector of directions
     pub fn new(directions: Vec<Direction>) -> Self {
         Path(directions)
     }
 
+    /// Get the length of the path
     pub fn len(&self) -> usize {
         self.0.len()
     }
 
+    /// Add a direction to the path
     pub fn push(&mut self, direction: Direction) {
         self.0.push(direction);
     }
 
+    /// Create a new random Path instance with a given length
     pub fn random<R: RngCore + Rng>(rng: &mut R) -> Self {
         Path((0..D).map(|_| rng.gen_range(0..2).into()).collect())
     }
 
+    /// Check if the path is empty
     pub fn is_empty(&self) -> bool {
         self.0.is_empty()
     }
@@ -164,13 +191,16 @@ impl From<usize> for Path {
 }
 
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+/// A block of data, represented as a vector of bytes
 pub struct Block(pub Vec<u8>);
 
 impl Block {
+    /// Create a new Block instance with a given vector of bytes
     pub fn new(data: Vec<u8>) -> Self {
         Block(data)
     }
 
+    /// Create a new random Block instance with a given size
     pub fn new_random() -> Self {
         let mut rng = ChaCha20Rng::from_entropy(); // Use ChaCha20Rng
         let mut block = vec![0u8; BLOCK_SIZE];
@@ -180,9 +210,11 @@ impl Block {
 }
 
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Default, Serialize, Deserialize)]
+/// A bucket of blocks, represented as a vector of Blocks
 pub struct Bucket(Vec<Block>);
 
 impl TreeValue for Bucket {
+    /// Create a new random Bucket instance with a given size
     fn new_random() -> Self {
         Bucket(vec![Block::new_random(); Z])
     }
@@ -211,39 +243,48 @@ impl IndexMut<usize> for Bucket {
 }
 
 impl Bucket {
+    /// Check if the bucket is empty
     pub fn is_empty(&self) -> bool {
         self.0.is_empty()
     }
 
+    /// Get the number of blocks in the bucket
     pub fn len(&self) -> usize {
         self.0.len()
     }
 
+    /// Add a block to the bucket
     pub fn push(&mut self, block: Block) {
         self.0.push(block);
     }
 
+    /// Get the block at a specific index
     pub fn get(&self, index: usize) -> Option<&Block> {
         self.0.get(index)
     }
 
+    /// Shuffle the blocks in the bucket using a random number generator
     pub fn shuffle<R: RngCore + Rng>(&mut self, rng: &mut R) {
         self.0.shuffle(rng);
     }
 
+    /// Get an iterator over the blocks in the bucket
     pub fn iter(&self) -> std::slice::Iter<'_, Block> {
         self.0.iter()
     }
 }
 
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+/// A cryptographic key, represented as a vector of bytes
 pub struct Key(pub Vec<u8>);
 
 impl Key {
+    /// Create a new Key instance with a given vector of bytes
     pub fn new(bytes: Vec<u8>) -> Key {
         Key(bytes)
     }
 
+    /// Create a new random Key instance with a given size
     pub fn random<R: RngCore + Rng>(rng: &mut R) -> Key {
         Key((0..LAMBDA / 8).map(|_| rng.gen()).collect())
     }
