+++ src/tree.rs
@@ -1,3 +1,10 @@
+//! This module provides binary tree implementations for both dense and sparse trees.
+//! 
+//! The main types are:
+//! - `BinaryTree<T>`: A dense binary tree implementation that stores values of type T
+//! - `SparseBinaryTree<T>`: A sparse binary tree that only stores non-empty nodes
+//! - `TreeValue`: A trait for values that can be stored in binary trees
+
 use std::{
     cmp::max,
     fmt::{self, Debug},
@@ -6,33 +13,46 @@ use std::{
     time::Instant,
 };
 
-use aes_gcm::aead::Buffer;
 use serde::{Deserialize, Serialize};
 
-use crate::{Bucket, Direction, Metadata, Path};
+use crate::{Bucket, Metadata, Path};
 
+/// A binary tree implementation that stores values of type T.
+/// 
+/// The tree is stored as a vector where:
+/// - Index 0 is unused
+/// - Index 1 is the root node
+/// - For any node at index i:
+///   - Left child is at index 2i
+///   - Right child is at index 2i + 1
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct BinaryTree<T> {
+    /// Vector storing the tree nodes, with None representing empty nodes
     pub value: Vec<Option<T>>,
 }
 
+/// A trait for values that can be stored in binary trees.
 pub trait TreeValue: Clone + Debug + PartialEq + Default {
+    /// Creates a new random value of this type
     fn new_random() -> Self;
 }
 
 impl<T: TreeValue> BinaryTree<T> {
+    /// Creates a new binary tree with a single value at the root
     pub fn new(value: T) -> Self {
         BinaryTree {
             value: vec![None, Some(value)],
         }
     }
 
+    /// Creates a new empty binary tree with space for root and one level
     pub fn new_empty() -> Self {
         BinaryTree {
             value: vec![None; 2],
         }
     }
 
+    /// Creates a new empty binary tree with space for the specified depth
     pub fn new_with_depth(depth: usize) -> Self {
         BinaryTree {
             // Prev 1 << depth + 1
@@ -42,10 +62,12 @@ impl<T: TreeValue> BinaryTree<T> {
         }
     }
 
+    /// Fills all nodes in the tree with the given value
     pub fn fill(&mut self, value: T) {
         self.value[1..].fill(Some(value));
     }
 
+    /// Inserts values along a path in the tree
     pub fn insert_path(&mut self, path: Path, values: Vec<T>) {
         let mut idx: usize = 1;
         self.value[1] = Some(values[0].clone());
@@ -59,6 +81,7 @@ impl<T: TreeValue> BinaryTree<T> {
         }
     }
 
+    /// Creates a tree from a vector of (values, path) pairs
     pub fn from_vec_with_paths(items: Vec<(Vec<T>, Path)>) -> Self
     where
         T: TreeValue,
@@ -70,10 +93,12 @@ impl<T: TreeValue> BinaryTree<T> {
         tree
     }
 
+    /// Returns the height of the tree
     pub fn height(&self) -> usize {
         ((self.value.len() as f64).log2().ceil() as usize) - 1
     }
 
+    /// Creates a tree from arrays of values and their indices
     pub fn from_array(values: Vec<T>, indices: Vec<usize>) -> Self {
         let mut tree = BinaryTree::new_empty();
         for (value, index) in values.iter().zip(indices) {
@@ -85,6 +110,7 @@ impl<T: TreeValue> BinaryTree<T> {
         tree
     }
 
+    /// Gets the value at a given path
     pub fn get(&self, path: &Path) -> Option<T> {
         let mut current = self.value[1].clone();
         let mut idx = 1;
@@ -98,6 +124,7 @@ impl<T: TreeValue> BinaryTree<T> {
         current
     }
 
+    /// Gets the index for a given path
     pub fn get_index(&self, path: &Path) -> usize {
         let mut current = 1;
         let mut idx = 1;
@@ -111,6 +138,7 @@ impl<T: TreeValue> BinaryTree<T> {
         idx
     }
 
+    /// Gets all nodes along a given path
     pub fn get_all_nodes_along_path(&self, path: &Path) -> Vec<T> {
         let start = Instant::now();
 
@@ -134,6 +162,7 @@ impl<T: TreeValue> BinaryTree<T> {
         nodes
     }
 
+    /// Finds the lowest common ancestor for a given path
     pub fn lca(&mut self, path: &Path) -> Option<(&mut T, Path)> {
         let mut current_path = Path::new(Vec::new());
         let mut idx = 1;
@@ -150,6 +179,7 @@ impl<T: TreeValue> BinaryTree<T> {
         self.value[idx].as_mut().map(|value| (value, current_path))
     }
 
+    /// Writes a value at a given path
     pub fn write(&mut self, value: T, path: Path) {
         let mut idx = 1;
         for direction in path {
@@ -162,6 +192,7 @@ impl<T: TreeValue> BinaryTree<T> {
         self.value[idx] = Some(value);
     }
 
+    /// Overwrites this tree with values from another tree
     pub fn overwrite(&mut self, other: &BinaryTree<T>) {
         if self.value.len() < other.value.len() {
             self.value.resize(other.value.len(), None);
@@ -176,6 +207,7 @@ impl<T: TreeValue> BinaryTree<T> {
             });
     }
 
+    /// Overwrites this tree with values from a sparse tree
     pub fn overwrite_from_sparse(&mut self, sparse_tree: &SparseBinaryTree<T>) {
         // Ensure that the binary tree has enough capacity to hold the elements from the sparse tree
         if let Some(&max_index) = sparse_tree.packed_indices.iter().max() {
@@ -194,6 +226,7 @@ impl<T: TreeValue> BinaryTree<T> {
         }
     }
 
+    /// Zips this tree with another tree, returning tuples of values and paths
     pub fn zip<S: Clone>(&self, rhs: &BinaryTree<S>) -> Vec<(Option<T>, Option<S>, Path)> {
         let len = max(self.value.len(), rhs.value.len());
         let mut lhs = self.value.clone();
@@ -215,6 +248,7 @@ impl<T: TreeValue> BinaryTree<T> {
             .collect()
     }
 
+    /// Zips this tree with another tree, returning tuples of mutable references and paths
     pub fn zip_mut<'a, 'b, S>(
         &'a mut self,
         rhs: &'b mut BinaryTree<S>,
@@ -258,17 +292,25 @@ impl fmt::Display for BinaryTree<Bucket> {
 
 // Sparse binary tree
 
+/// A sparse binary tree implementation that only stores non-empty nodes.
+/// 
+/// Instead of storing all nodes in a vector like BinaryTree, this implementation
+/// only stores the non-empty nodes in a compressed format using:
+/// - packed_buckets: Vector of actual values
+/// - packed_indices: Vector of indices where those values belong
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct SparseBinaryTree<T> {
-    pub packed_buckets: Vec<T>,     // List of non-None buckets
-    pub packed_indices: Vec<usize>, // Indices corresponding to these buckets
+    /// List of non-None buckets
+    pub packed_buckets: Vec<T>,
+    /// Indices corresponding to these buckets
+    pub packed_indices: Vec<usize>,
 }
 
 impl<T> SparseBinaryTree<T>
 where
     T: Clone + Debug + PartialEq + Default,
 {
-    // Creates a new empty sparse binary tree
+    /// Creates a new empty sparse binary tree
     pub fn new() -> Self {
         SparseBinaryTree {
             packed_buckets: Vec::new(),
@@ -276,7 +318,7 @@ where
         }
     }
 
-    // Creates a new sparse binary tree with provided packed_buckets and packed_indices
+    /// Creates a new sparse binary tree with provided packed_buckets and packed_indices
     pub fn new_with_data(packed_buckets: Vec<T>, packed_indices: Vec<usize>) -> Self {
         SparseBinaryTree {
             packed_buckets,
@@ -297,7 +339,7 @@ where
         }
     }
 
-    // Retrieve the value at the given path in a sparse binary tree
+    /// Retrieve the value at the given path in a sparse binary tree
     pub fn get(&self, path: &Path) -> Option<&T> {
         let mut idx = 1; // Start at the root
         for &direction in path {
@@ -306,7 +348,7 @@ where
         self.get_by_index(idx)
     }
 
-    // Retrieve the value at the given path in a sparse binary tree
+    /// Get the index for a given path
     pub fn get_index(&self, path: &Path) -> usize {
         let mut idx = 1; // Start at the root
         for &direction in path {
@@ -315,7 +357,7 @@ where
         idx
     }
 
-    // Retrieves value by index
+    /// Retrieves value by index
     pub fn get_by_index(&self, index: usize) -> Option<&T> {
         self.packed_indices
             .iter()
@@ -323,7 +365,7 @@ where
             .map(|pos| &self.packed_buckets[pos])
     }
 
-    // Retreives a reference to the bucket by index
+    /// Retrieves a mutable reference to the bucket by index
     pub fn get_by_index_mut(&mut self, index: usize) -> Option<&mut T> {
         self.packed_indices
             .iter()
@@ -331,7 +373,7 @@ where
             .map(move |pos| &mut self.packed_buckets[pos]) // Returns mutable reference
     }
 
-    // Write a value into the sparse tree at the specified path
+    /// Write a value into the sparse tree at the specified path
     pub fn write(&mut self, value: T, path: Path) {
         let mut idx = 1; // Start at the root
 
@@ -342,7 +384,7 @@ where
         self.add_bucket(idx, value);
     }
 
-    // Find the lowest common ancestor (LCA) of a given path
+    /// Find the lowest common ancestor (LCA) index of a given path
     pub fn lca_idx(&self, path: &Path) -> Option<(usize, Path)> {
         let mut current_path = Path::new(Vec::new());
         let mut idx = 1; // Start at the root
@@ -358,7 +400,7 @@ where
         Some((idx, current_path.clone()))
     }
 
-    // Find the lowest common ancestor (LCA) of a given path
+    /// Find the lowest common ancestor (LCA) of a given path
     pub fn lca(&mut self, path: &Path) -> Option<(&mut T, Path)> {
         let mut current_path = Path::new(Vec::new());
         let mut idx = 1; // Start at the root
@@ -378,14 +420,7 @@ where
             .map(|value| (value, current_path.clone()))
     }
 
-    // // Overwrite the sparse binary tree with another one
-    // pub fn overwrite(&mut self, other: &SparseBinaryTree<T>) {
-    //     for (index, bucket) in other.indices.iter().zip(&other.packed_buckets) {
-    //         self.add_bucket(*index, bucket.clone());
-    //     }
-    // }
-
-    // Zips two sparse binary trees together
+    /// Zips two sparse binary trees together
     pub fn zip<S: Clone>(&self, rhs: &SparseBinaryTree<S>) -> Vec<(Option<T>, Option<S>, Path)> {
         let mut results = Vec::new();
 
@@ -404,6 +439,7 @@ where
         results
     }
 
+    /// Returns an iterator that zips two sparse binary trees together with mutable references
     pub fn zip_mut<'a, S>(
         &'a mut self,
         rhs: &'a mut SparseBinaryTree<S>,
@@ -411,6 +447,7 @@ where
         ZipMutIterator::new(self, rhs)
     }
 
+    /// Zips this sparse tree with a regular binary tree
     pub fn zip_with_binary_tree<S: Clone>(
         &self,
         rhs: &BinaryTree<S>,
@@ -438,6 +475,7 @@ where
         results
     }
 
+    /// Gets all nodes along a given path
     pub fn get_all_nodes_along_path(&self, path: &Path) -> Vec<&T> {
         let mut nodes = Vec::new();
         let mut idx = 1;  // Start at root
@@ -467,6 +505,7 @@ struct SlicePair<'a, T, S> {
     right_buckets: &'a mut [S],
 }
 
+/// Iterator for zipping two sparse binary trees together with mutable references
 pub struct ZipMutIterator<'a, T, S> {
     slices: Option<SlicePair<'a, T, S>>,
 }
@@ -525,7 +564,7 @@ impl<'a, T, S> Iterator for ZipMutIterator<'a, T, S> {
     }
 }
 
-
+/// State of the database trees
 #[derive(Serialize, Deserialize)]
 struct DBState {
     tree: BinaryTree<Bucket>,
@@ -534,10 +573,15 @@ struct DBState {
 
 /// Parameters that define the state of the DB.
 pub struct DBStateParams {
+    /// Size of the buckets in the tree
     pub bucket_size: usize,
+    /// Number of iterations in the tree
     pub num_iters: usize,
+    /// Depth of the tree
     pub depth: usize,
+    /// Number of clients in the tree
     pub num_clients: usize,
+    /// Timestamp of the tree
     pub timestamp: u64,
 }
 
@@ -578,4 +622,3 @@ pub fn deserialize_trees(params: &DBStateParams) -> (BinaryTree<Bucket>, BinaryT
 
     (db_state.tree, db_state.metadata)
 }
-
